requirements.txt
PyMuPDF==1.27.1
pytesseract==0.3.13
Pillow==12.1.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
google-genai==1.65.0
numpy==1.26.4
streamlit==1.55.0
